CP-25261: Fix Cluster_host.disable prerequisites
The prerequisites for Cluster_host.disable should check that there
are no attached PBDs on the host associated with an SR whose SM record
requires the cluster stack of the cluster. The code was previously
considering all PBDs, regardless of it being attached or not; this
commit fixes this.

The prerequisite-checking is also put into its own function to make
testing easier and to be callable from Cluster.destroy, which has the
same prerequisites for the last remaining cluster_host record.

Signed-off-by: Frederico Mazzone <frederico.mazzone@citrix.com>