CA-270443: Avoid deadlock in SR.create / PBD.plug
PBD.plug was trying to acquire the host-local clustering lock while SR.create
was still holding it, leading to a deadlock.

Only protect the body of SR create with the host-local clustering lock,
and release the lock before plugging the PBD.
The lock is meant to protect against disabling clustering while you have an SR
using it and to avoid race conditions between creating and destroying an SR and
a cluster. With the new lock scope that is still the case: we only release the
lock at the very end - after we've created the SR - when plugging each PBD.

But those PBD plugs need to be protected by locks and allowed-operations on their own anyway,
because you can unplug and replug PBDs many times.
Therefore this change shouldn't introduce race conditions.

Signed-off-by: Edwin Török <edvin.torok@citrix.com>
Signed-off-by: Frederico Mazzone <frederico.mazzone@citrix.com>