CA-274585: allow unplugging statefile VDI
If the HA daemon is not running it is safe to unplug the statefile VDI.
Need to allow unplugging the statefile VDI if we want to unplug all PBDs on
shutdown.

Note: this fixes shutdown on HA slaves, but not on HA pool master.
There the metadata VDI will stay plugged: CA-276993.

Signed-off-by: Edwin Török <edvin.torok@citrix.com>