CA-274585: unplug all local PBDs on shutdown/reboot
We need to unplug the PBDs before shutdown/reboot, while we still have
functional networking.

We should not rely only on systemd to unmount the filesystems, because
SM would also expect PBD.unplug to get called on clean shutdown/reboot.
For some SRs (like GFS2) we also need to perform additional cleanup operations
that can only be done after all the SRs are unplugged.

The unplug must be done after HA is disabled, otherwise the statefile might
still be using it.

There are 3 ways to shutdown:
 * stop xapi-domains.service, xe host-disable, xe host-shutdown
 * stop from XC which shuts down VMs, and then does the host shutdown
 * xe host-shutdown directly with VMs running (not prevented, but the docs say you shouldn't do this)

Call the unplug in both places, this is a best effort unplug, i.e.
unplug as many PBDs as we can and ignore errors.

Signed-off-by: Edwin Török <edvin.torok@citrix.com>