[CA-273985] Take clustering lock only if SM requires a cluster stack
Currently in xapi_pbd and xapi_sr we take a clustering lock (`with_clustering_lock`) regardless of the SR type.
However we only need to do this if the SM type requires a cluster stack. This commit adds a new function `with_clustering_lock_if_needed`, which calls `with_clustering_lock` if the SM type requires a cluster stack, otherwise it executes the passed function without any wrapper. The calls in xapi_pdb and xapi_sr are changed to use this new function.

This is tested by the fact that nesting two calls to `with_clustering_lock` will deadlock. We test that the lock is correctly taken when needed by nesting two calls to `with_clustering_lock_if_needed`, which should both take the lock and therefore deadlock. If they don't then the test fails, otherwise it times out and passes. We test hat the lock is not taken when it's not needed by nesting two calls similarly, but this time one requires a cluster stack and the other doesn't. This shouldn't deadlock, which will cause the test the timeout and fail.

Signed-off-by: Callum McIntyre <callum.mcintyre@citrix.com>