[CP-25971] Introduce Cluster.pool_force_destroy (#3423)
* CP-25971: draft Cluster force destroy API

Signed-off-by: Edwin Török <edvin.torok@citrix.com>

* [CP-25971] Rename forget to force_destroy

Signed-off-by: Callum McIntyre callum.mcintyre@citrix.com

* [CP-25971] Add Cluster.pool_force_destroy

Cluster.pool_force_destroy is a best-effort function that performs as much Cluster cleanup as it can given the current state of the cluster. It is assumed that it typically will not complete successfully, because in the case that it can it would be better to call `Cluster.pool_destroy` instead. Instead it completes as much as it can, as cleanly as it can, and aims to give a useful error message if it cannot complete.

It first tries to cleanly disable (ie leave the cluster) each slave. We expect exceptions here, and ignore them. Next we try to force disable each host, including the master. We include the master here because if it's the only healthy node then we can at least remove it (it can't cleanly leave because it's the last node). This time we collect exceptions instead of ignoring them - but we attempt to force destroy all nodes before raising any error so we clean up everything that we possibly can. If there are no errors here then we destroy the cluster, but if a force destroy failed then we report that failure to the user and stop.

This commit also modifies `Cluster.destroy` to support clusters with 0 or 1 nodes, needed in the case where we managed to clean up all nodes, and makes `Cluster.pool_destroy` support the case where the pool master is Cluster_host is down in the same way that `Cluster.pool_force_destroy` does.

Tested scenarios:
- In a cluster of 3 nodes where 2 are down (so there is no quorum, and the up node can't leave), we clean up the active node successfully.
- In a cluster of 3 with all nodes available, we clean up all nodes and destroy the cluster successfully.

Signed-off-by: Callum McIntyre callum.mcintyre@citrix.com